Move alibaba-druid into jdbc folder
