+++ agent/agent-plugins/jdbc/alibaba-druid/src/main/java/org/bithon/agent/plugin/jdbc/alibaba/druid/interceptor/DruidDataSource$GetValueAndReset.java
@@ -14,15 +14,15 @@
  *    limitations under the License.
  */
 
-package org.bithon.agent.plugin.alibaba.druid.interceptor;
+package org.bithon.agent.plugin.jdbc.alibaba.druid.interceptor;
 
 import com.alibaba.druid.pool.DruidDataSource;
 import com.alibaba.druid.pool.DruidDataSourceStatValue;
 import org.bithon.agent.instrumentation.aop.context.AopContext;
 import org.bithon.agent.instrumentation.aop.interceptor.declaration.AfterInterceptor;
 import org.bithon.agent.observability.metric.domain.jdbc.JdbcPoolMetrics;
-import org.bithon.agent.plugin.alibaba.druid.metric.MonitoredSource;
-import org.bithon.agent.plugin.alibaba.druid.metric.MonitoredSourceManager;
+import org.bithon.agent.plugin.jdbc.alibaba.druid.metric.MonitoredSource;
+import org.bithon.agent.plugin.jdbc.alibaba.druid.metric.MonitoredSourceManager;
 
 /**
  * {@link com.alibaba.druid.pool.DruidDataSource#getStatValueAndReset()}
