+++ agent/agent-plugins/jdbc/alibaba-druid/src/main/java/org/bithon/agent/plugin/jdbc/alibaba/druid/metric/MonitoredSourceManager.java
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 
-package org.bithon.agent.plugin.alibaba.druid.metric;
+package org.bithon.agent.plugin.jdbc.alibaba.druid.metric;
 
 import com.alibaba.druid.pool.DruidDataSource;
 import org.bithon.agent.observability.utils.MiscUtils;
